:art: refactor: Use Array.from where iterable spread is weird for style.
This is not enough to make microbundle's output correct. But we are
ditching microbundle in the next commit anyway. Probably replacing the
spread operators with `Array.from` in the following lines would be enough:

```
const expectedLengths = Array.from(
	new Set([...validLengths, ...validLengthsWithTrailingSlash]),
);
```

But I can't tolerate the output of a bundler being incorrect. So I won't
change this now.